requirements.txt
paho-mqtt
bluepy
pi_sht1x
RPi.GPIO